Reorganization and README
