+++ .gitignore
@@ -1,7 +1,7 @@
 node_modules
 public/bundle.js
 public/bundle.js.map
-secrets.js
+secrets
 .DS_Store
 npm-debug.log
 yarn-error.log
