+++ README.md
@@ -1,1 +1,10 @@
-# WORKING TITLE README
+# Gathr.io
+
+## Developers:
+
+- Steps for Installation
+  - ```git clone https://github.com/Gathr-io/Gathr.io.git``` to clonethe repository to your machine
+  - ```npm install``` node modules
+  - On your terminal ```createDb Gathr``` to create a test database
+  - ```npm run seed``` to seed the database
+  - ```npm run start-dev``` for testing on [your local host](http://localhost:8080)
