+++ server/db/db.js
@@ -1,6 +1,6 @@
 const Sequelize = require('sequelize')
 const db = new Sequelize(
-  process.env.DATABASE_URL || 'postgres://localhost:5432/boilermaker', {
+  process.env.DATABASE_URL || 'postgres://localhost:5432/Gathr', {
     logging: false
   }
 )
