+++ server/index.js
@@ -21,7 +21,7 @@ module.exports = app
  * keys as environment variables, so that they can still be read by the
  * Node process on process.env
  */
-if (process.env.NODE_ENV !== 'production') require('../secrets')
+if (process.env.NODE_ENV !== 'production') require('../secrets/secrets.js')
 
 // passport registration
 passport.serializeUser((user, done) => done(null, user.id))
